# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"APPLICATION_STATUS": {
				"enum": [
					"INITIALIZING",
					"ERROR",
					"OK"
				],
				"type": "string"
			},
			"CredentialStatus": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"StatusList2021Entry"
						],
						"nullable": false
					},
					"statusPurpose": {
						"type": "string",
						"enum": [
							"revocation",
							"suspension"
						]
					},
					"statusListIndex": {
						"type": "string"
					},
					"statusListCredential": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"statusPurpose",
					"statusListIndex",
					"statusListCredential"
				],
				"type": "object",
				"additionalProperties": true
			},
			"JSONObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"$ref": "#/components/schemas/JSONValue"
				}
			},
			"JSONArray": {
				"properties": {},
				"type": "object",
				"additionalProperties": true
			},
			"JSONValue": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					},
					{
						"$ref": "#/components/schemas/JSONObject"
					},
					{
						"$ref": "#/components/schemas/JSONArray"
					}
				]
			},
			"SignedVerifiableCredential": {
				"properties": {
					"@context": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"$ref": "#/components/schemas/JSONObject"
								}
							]
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"issuer": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						]
					},
					"issuanceDate": {
						"type": "string"
					},
					"expirationDate": {
						"type": "string"
					},
					"credentialSubject": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/JSONObject"
							},
							{
								"$ref": "#/components/schemas/JSONValue"
							}
						]
					},
					"credentialStatus": {
						"$ref": "#/components/schemas/CredentialStatus"
					},
					"proof": {
						"properties": {
							"proofValue": {
								"type": "string"
							},
							"jws": {
								"type": "string"
							},
							"proofPurpose": {
								"type": "string"
							},
							"verificationMethod": {
								"type": "string"
							},
							"nonce": {
								"type": "string"
							},
							"domain": {
								"type": "string"
							},
							"challenge": {
								"type": "string"
							},
							"created": {
								"type": "string"
							},
							"type": {
								"type": "string"
							}
						},
						"required": [
							"jws",
							"proofPurpose",
							"verificationMethod",
							"created",
							"type"
						],
						"type": "object"
					}
				},
				"required": [
					"@context",
					"type",
					"issuer",
					"issuanceDate",
					"credentialSubject",
					"proof"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Credential": {
				"properties": {
					"@context": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"$ref": "#/components/schemas/JSONObject"
								}
							]
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"issuer": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						]
					},
					"issuanceDate": {
						"type": "string"
					},
					"expirationDate": {
						"type": "string"
					},
					"credentialSubject": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/JSONObject"
							},
							{
								"$ref": "#/components/schemas/JSONValue"
							}
						]
					},
					"credentialStatus": {
						"$ref": "#/components/schemas/CredentialStatus"
					}
				},
				"required": [
					"@context",
					"type",
					"issuer",
					"issuanceDate",
					"credentialSubject"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CredentialIssueOptions": {
				"properties": {
					"created": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"credentialStatus": {
						"$ref": "#/components/schemas/CredentialStatus"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"CredentialIssueParams": {
				"properties": {
					"credential": {
						"$ref": "#/components/schemas/Credential"
					},
					"options": {
						"$ref": "#/components/schemas/CredentialIssueOptions"
					}
				},
				"required": [
					"credential"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CredentialVerifyOptions": {
				"properties": {
					"challenge": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"credentialStatus": {
						"properties": {
							"type": {
								"type": "string",
								"enum": [
									"StatusList2021Entry"
								],
								"nullable": false
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"CredentialVerifyParams": {
				"properties": {
					"verifiableCredential": {
						"$ref": "#/components/schemas/SignedVerifiableCredential"
					},
					"options": {
						"$ref": "#/components/schemas/CredentialVerifyOptions"
					}
				},
				"required": [
					"verifiableCredential"
				],
				"type": "object",
				"additionalProperties": true
			},
			"StatusList2021Credential": {
				"properties": {
					"@context": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"issuer": {
						"type": "string"
					},
					"issuanceDate": {
						"type": "string"
					},
					"credentialSubject": {
						"properties": {
							"encodedList": {
								"type": "string"
							},
							"statusPurpose": {
								"type": "string",
								"enum": [
									"revocation"
								],
								"nullable": false
							},
							"type": {
								"type": "string",
								"enum": [
									"StatusList2021"
								],
								"nullable": false
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"encodedList",
							"statusPurpose",
							"type",
							"id"
						],
						"type": "object"
					},
					"proof": {
						"properties": {
							"jws": {
								"type": "string"
							},
							"proofPurpose": {
								"type": "string",
								"enum": [
									"assertionMethod"
								],
								"nullable": false
							},
							"verificationMethod": {
								"type": "string"
							},
							"created": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"Ed25519Signature2018"
								],
								"nullable": false
							}
						},
						"required": [
							"jws",
							"proofPurpose",
							"verificationMethod",
							"created",
							"type"
						],
						"type": "object"
					}
				},
				"required": [
					"@context",
					"id",
					"type",
					"issuer",
					"issuanceDate",
					"credentialSubject"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateCredentialStatusParams": {
				"properties": {
					"credentialId": {
						"type": "string"
					},
					"credentialStatus": {
						"items": {
							"properties": {
								"status": {
									"type": "string",
									"enum": [
										"true",
										"false"
									]
								},
								"type": {
									"type": "string",
									"enum": [
										"revocation"
									],
									"nullable": false
								}
							},
							"required": [
								"status",
								"type"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"credentialId",
					"credentialStatus"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DidDocument": {
				"type": "string"
			},
			"SignedPresentation": {
				"properties": {
					"@context": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"$ref": "#/components/schemas/JSONObject"
								}
							]
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"holder": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/JSONObject"
							}
						]
					},
					"verifiableCredential": {
						"items": {
							"$ref": "#/components/schemas/SignedVerifiableCredential"
						},
						"type": "array"
					},
					"proof": {
						"properties": {
							"jws": {
								"type": "string"
							},
							"verificationMethod": {
								"type": "string"
							},
							"proofPurpose": {
								"type": "string"
							},
							"challenge": {
								"type": "string"
							},
							"created": {
								"type": "string"
							},
							"type": {
								"type": "string"
							}
						},
						"required": [
							"jws",
							"verificationMethod",
							"proofPurpose",
							"created",
							"type"
						],
						"type": "object"
					}
				},
				"required": [
					"@context",
					"type",
					"holder",
					"verifiableCredential",
					"proof"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Presentation": {
				"properties": {
					"@context": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"$ref": "#/components/schemas/JSONObject"
								}
							]
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"holder": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/JSONObject"
							}
						]
					},
					"verifiableCredential": {
						"items": {
							"$ref": "#/components/schemas/SignedVerifiableCredential"
						},
						"type": "array"
					}
				},
				"required": [
					"@context",
					"type",
					"holder",
					"verifiableCredential"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ProvePresentationParams": {
				"properties": {
					"presentation": {
						"$ref": "#/components/schemas/Presentation"
					},
					"options": {
						"properties": {
							"challenge": {
								"type": "string"
							}
						},
						"required": [
							"challenge"
						],
						"type": "object"
					}
				},
				"required": [
					"presentation",
					"options"
				],
				"type": "object",
				"additionalProperties": true
			},
			"VerifyPresentationParams": {
				"properties": {
					"verifiablePresentation": {
						"$ref": "#/components/schemas/SignedPresentation"
					},
					"options": {
						"properties": {
							"challenge": {
								"type": "string"
							}
						},
						"required": [
							"challenge"
						],
						"type": "object"
					}
				},
				"required": [
					"verifiablePresentation",
					"options"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "hedera-vc-api",
		"version": "0.0.1",
		"description": "REST API to manage Hedera Verifiable Credentials, Presentations and DIDs",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Olivier Scherrer ",
			"email": "olivier.scherrer@meranti.fr"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/init": {
			"post": {
				"operationId": "CreateIdentityNetwork",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"statusListFileId": {
											"type": "string"
										},
										"didDocument": {
											"type": "string"
										}
									},
									"required": [
										"statusListFileId",
										"didDocument"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/admin/status": {
			"get": {
				"operationId": "GetVerboseStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/APPLICATION_STATUS"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/credentials/issue": {
			"post": {
				"operationId": "IssueCredential",
				"responses": {
					"201": {
						"description": "Issued",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignedVerifiableCredential"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CredentialIssueParams"
							}
						}
					}
				}
			}
		},
		"/credentials/verify": {
			"post": {
				"operationId": "VerifyCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"items": {},
											"type": "array"
										},
										"warnings": {
											"items": {},
											"type": "array"
										},
										"checks": {
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									},
									"required": [
										"errors",
										"warnings",
										"checks"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CredentialVerifyParams"
							}
						}
					}
				}
			}
		},
		"/credentials/status/{statusListId}": {
			"get": {
				"operationId": "GetStatusList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StatusList2021Credential"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "statusListId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SetStatusList",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "statusListId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCredentialStatusParams"
							}
						}
					}
				}
			}
		},
		"/did": {
			"get": {
				"operationId": "GetDidDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/presentations/prove": {
			"post": {
				"operationId": "ProvePresentation",
				"responses": {
					"201": {
						"description": "Proved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignedPresentation"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProvePresentationParams"
							}
						}
					}
				}
			}
		},
		"/presentations/verify": {
			"post": {
				"operationId": "VerifyPresentation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"items": {},
											"type": "array"
										},
										"warnings": {
											"items": {},
											"type": "array"
										},
										"checks": {
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									},
									"required": [
										"errors",
										"warnings",
										"checks"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyPresentationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}